Create doge.js
doge.js is generates 1000 Dogecoin wallet addresses and private keys using the dogecore-lib library.

It starts by requiring the dogecore-lib and fs libraries. The dogecore-lib library provides functions for generating and manipulating Dogecoin addresses and private keys, while the fs (file system) library provides functions for reading and writing files.

Next, it uses a for loop to generate 1000 wallet addresses and private keys using the dogecore.PrivateKey.fromRandom() function. This function generates a new random private key and returns a PrivateKey object.

For each PrivateKey object, it then calls the toAddress() function to get the corresponding wallet address and the toWIF() function to get the private key in the Wallet Import Format (WIF). These are added to an object, which is then pushed onto the wallet's array.

Finally, it writes the wallet addresses and private keys to two separate files, addr.txt and priv.txt, using the writeFileSync function from the fs library. The wallet addresses are written to addr.txt and the private keys are written to priv.txt. The map function is used to extract the relevant data from each object in the wallets array, and the join function is used to concatenate the resulting arrays into strings with newline characters between each item.